Add load library to plugins-utils.h
